+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1.  UI Toolkit: `MinMaxSlider` friendly error if been used on wrong type [#232](https://github.com/TylerTemp/SaintsField/issues/232)
 2.  UI Toolkit: `ShowInInspector` now shows an error box if the target attributes raises an error in its getter
+3.  UI Toolkit: `ShowInInspector` now always update the sub-field value display if the target (e.g. a class) has sub-fields/properties
 
 ## 4.9.0 ##